modify Product component
